Add unittest
