+++ test_email_validation.py
@@ -1,5 +1,5 @@
 
-import re
+import email_validation
 import unittest
 
 
@@ -7,14 +7,6 @@ class EmailValidationTest(unittest.TestCase):
     """unit test for EmailValidation"""
     def test_validate_email(self):
         self.assertTrue(e_mail, validate_email())
-
-
-def validate_email(a):
-    pattern = r'^([a-z]+|[0-9])*@[a-z]+_?[a-z]+\.com$'
-    if re.match(pattern, a):
-        print('True')
-    else:
-        print('False')
 
 
 e_mail = 'vasya@gde-to.ru'
